+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - Adding AmiById Policy
   [#89](https://github.com/pulumi/pulumi-policy-aws/pull/89)
-  
+- Added in checks for `alb` and `lb` to the compute and network policies
+
 ---
 
 ## 0.4.0 (2022-09-22)